+++ src/App.jsx
@@ -10,6 +10,7 @@ import RoutePoints from "./components/routes/RoutePoints";
 import PersonalInformationPolicyPage from "./components/PersonalInformationPolicyPage";
 import Navbar from "./components/Navbar";
 import FAQ from "./components/FAQ";
+import SeeOnMap from "./components/SeeOnMap";
 
 function App() {
   const [selectedRoute, setSelectedRoute] = useState(null);
@@ -18,69 +19,32 @@ function App() {
   const [openStreetMapConsent, setOpenStreetMapConsent] = useState(null);
   const [lat, setLat] = useState(null);
   const [long, setLong] = useState(null);
-  const [userAllowedAccessToGeoLocation, setUserAllowedAccessToGeoLocation] = useState(false);
+  const locationUpdateInterval = 30000;
 
   const contextLatLong = useMemo(
     () => ({
       lat,
       long,
     }),
-    [lat, long],
+    [lat, long]
   );
 
-  const updateLocation = () => {
-    if (lat === null || long === null) {
+  function startLocationPrompter() {
+    setInterval(() => {
       navigator.geolocation.getCurrentPosition((position) => {
         setLat(position.coords.latitude);
         setLong(position.coords.longitude);
       });
-    }
-    setTimeout(() => {
-      navigator.geolocation.getCurrentPosition((position) => {
-        setLat(position.coords.latitude);
-        setLong(position.coords.longitude);
-      });
-      updateLocation();
-    }, 3000);
-  };
-
-  const handlePermissions = async () => {
-    if (navigator.permissions?.query) {
-      // Not apple
-      const permissions = await navigator.permissions.query({
-        name: "geolocation",
-      });
-      const { state } = permissions;
-      permissions.onchange = (event) => {
-        if (event.target.state === "granted") {
-          setUserAllowedAccessToGeoLocation(true);
-        }
-        if (event.target.state === "denied") {
-          setUserAllowedAccessToGeoLocation(false);
-        }
-      };
-      if (state === "granted") {
-        updateLocation();
-        setUserAllowedAccessToGeoLocation(true);
-      } else if (state === "prompt") {
-        updateLocation();
-        setUserAllowedAccessToGeoLocation(false);
-      } else if (state === "denied") {
-        setUserAllowedAccessToGeoLocation(false);
-      }
-    } else {
-      // apple
-      updateLocation();
-    }
-  };
-
-  const requestPermissions = async () => {
-    handlePermissions();
-  };
+    }, locationUpdateInterval);
+  }
 
   useEffect(() => {
-    requestPermissions();
-  }, [userAllowedAccessToGeoLocation]);
+    navigator.geolocation.getCurrentPosition((position) => {
+      setLat(position.coords.latitude);
+      setLong(position.coords.longitude);
+    });
+    startLocationPrompter();
+  }, []);
 
   useEffect(() => {
     // Consent for handling data with regards to open street map
@@ -94,10 +58,7 @@ function App() {
     <div className="App flex flex-col h-full pt-24 min-h-screen dark:text-white w-screen pl-3 pr-3 pb-3 text-zinc-800 bg-zinc-100 dark:bg-zinc-800 overflow-hidden">
       <LatLongContext.Provider value={contextLatLong}>
         <PermissionContext.Provider
-          value={useMemo(
-            () => ({ userAllowedAccessToGeoLocation, openStreetMapConsent, setOpenStreetMapConsent }),
-            [userAllowedAccessToGeoLocation, openStreetMapConsent],
-          )}
+          value={useMemo(() => ({ openStreetMapConsent, setOpenStreetMapConsent }), [openStreetMapConsent])}
         >
           <RouteContext.Provider
             value={useMemo(
@@ -109,7 +70,7 @@ function App() {
                 listOfUnlocked,
                 setListOfUnlocked,
               }),
-              [selectedRoute, nextUnlockablePointId, listOfUnlocked],
+              [selectedRoute, nextUnlockablePointId, listOfUnlocked]
             )}
           >
             <Navbar />
@@ -132,6 +93,9 @@ function App() {
                 </Route>
                 <Route path="/info">
                   <Info />
+                </Route>
+                <Route path="/see-on-map/:latitude/:longitude">
+                  <SeeOnMap />
                 </Route>
                 <Route path="/">
                   <FrontPage />